requirements.txt
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
awscli==1.20.11
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
boto==2.49.0
boto3==1.19.2
botocore==1.22.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
crypto==1.4.1
cryptography==3.4.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.8
django-mysql==4.2.0
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.1.0
fastjsonschema==2.15.3
filelock==3.0.12
fonttools==4.33.3
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.4.2
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
mysql-connector==2.2.9
mysqlclient==2.0.3
Naked==0.1.31
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netaddr==0.8.0
nltk==3.7
notebook==6.4.5
numpy==1.20.3
oauthlib==3.2.0
opencv-python==4.5.3.56
openpyxl==3.0.7
packaging==21.0
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.21
psutil==5.8.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.20
pycryptodome==3.10.1
pyee==8.1.0
PyGetWindow==0.0.9
Pygments==2.10.0
PyMsgBox==1.0.9
pyparsing==3.0.0
pyperclip==1.8.2
pyppeteer==0.2.5
pyppeteer-stealth==2.7.3
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.27
python-anticaptcha==0.7.1
python-dateutil==2.8.2
python-decouple==3.5
PyTweening==1.0.3
pytz==2021.3
pywhatkit==5.0
pywin32==302
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
redis==3.5.3
regex==2022.4.24
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.5.0
schedule==1.1.0
scikit-learn==1.1.1
scipy==1.8.1
selenium==3.141.0
selenium-stealth==1.0.6
Send2Trash==1.8.0
sentry-sdk==1.1.0
shellescape==3.8.1
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
sqlparse==0.4.2
static3==0.7.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.60.0
traitlets==5.1.0
TwitterSearch==1.0.2
Unidecode==1.2.0
urllib3==1.26.7
virtualenv==20.4.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
xlrd==2.0.1
